+++ src/Symfony/Component/Form/FormRenderer.php
@@ -70,9 +70,10 @@ class FormRenderer implements FormRendererInterface
     /**
      * {@inheritdoc}
      */
-    public function setTheme(FormView $view, $themes)
+    public function setTheme(FormView $view, $themes /*, $useDefaultThemes = true */)
     {
-        $this->engine->setTheme($view, $themes);
+        $args = func_get_args();
+        $this->engine->setTheme($view, $themes, isset($args[2]) ? (bool) $args[2] : true);
     }
 
     /**